pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "careerpilot"
version = "0.1.0"
description = "Autonomous AI agent for internship preparation"
requires-python = ">=3.11"
dependencies = [
    "groq",
    "requests",
    "pydantic",
    "pyyaml",
    "loguru",
    "rich",
    "apscheduler",
    "sqlite-utils",
    "python-dotenv",
    "pytest",
]

[project.scripts]
careerpilot = "agent:run"